chore: add vibes to medications folder
